requirements.txt
requests
python-dotenv
cohere
colorama
streamlit